Rpi cam cutout almost finished
I think the Z tolerances aren't quite right; would be nice to
push the camera chip just enough that it's flat; at least on
my unit the foam thingy the camera is sitting on causes it to
be at a slight angle sometimes.

Other than that, the cutout looks good.
And the regular non-cutout model still works too.