Split vars out of RPcam2
Split the vars out of Rpcam2 so I can use the variables
via "include" and the objects via "use".
Not sure if this is the right thing to do, but "use" doesn't
pick up the variables, and "include" renders the test object,
so I don't see any other options at the moment.