Tombstone camera mount tweaks
The lens position is now about 10mm higher.
Also there's a little shield to cover the camera cable.
Not sure the shield will work out as-is, it might be impossible
to feed the cable through it due to the small size and square
angle at the bottom. Next version needs an angled slot at
the bottom.