First successful MP10 bed shim
Can now read the bed pretty well.
Wrote a second little script to analyze the data.
Did a little number crunching to support the idea of
shimming with a large sheet of paper with masking tape
on it. Built one, installed it.

Good news is the bed is much flatter and my test print with
two-layer squares mostly succeeds now, which is much better
than the total failure we had earlier.

Bad news is that shimming with paper almost certainly is
affecting the temperature of the bed; putting an insulator
under the metal plate can't be a good thing!

Experiments will continue.