Added LED ring and shroud
Tombstone camera mount now has a small shroud on the lens
to cut down glare from overhead lights.
Also added a spot to mount a small LED ring for front
illumination. Haven't tried it in practice yet but the
little Adafruit LED ring does fit, and there are spots
to run the wires.
Also added a little reinforcement to make it a little more
difficult to break by bumping it than the previous version.