New universal mounting bar
To attach to the little shelf on my desk.
First couple versions had round holes in the back which causes
significant support issues.

Current version uses square holes in the back. They overhang a
couple mm, but that should help keep the nuts in place. I hope.
If that doesn't work, next idea is to taper the inner holes so
they don't make a steep overhang but can still retain the nut.